Fixed resizing issue
Moved all panels from other classes (windowRightPanel, windowPanelLeft, etc.) into one single panel (found in window() method) which is then placed in the frame by itself so the panels will hopefully resize properly with the frame.

Also made changes to how the jtable is passed into the Frame class so that it will resize properly with the frame.